Move UNNotification & FirebaseMessaging into its own manager
